set up plain text plugin
