+++ src/components/Editor/Editor.css
@@ -1,0 +1,30 @@
+.editorContainer {
+  position: relative;
+}
+
+.editorInput {
+  border: 1px solid #ccc;
+  padding: 10px 8px;
+  border-radius: 5px;
+  min-height: 100px;
+}
+.editorInput * {
+  margin: 0;
+}
+
+.editorPlaceholder {
+  position: absolute;
+  top: 10px;
+  left: 11px; /* add some extra space so cursor doesn't overlap with placeholder  */
+  color: #999;
+  pointer-events: none;
+}
+
+/* default theme */
+.defaultBold {
+  font-weight: bold;
+}
+
+.defaultItalic {
+  font-style: italic;
+}
