+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from "./components/Editor";
+export * from "./components/Editor/Editor";
